+++ src-webviews/src/pages/components.ts
@@ -8,10 +8,12 @@ import InputBox from './input/InputBox.vue';
 import Job from './job/Job.vue';
 import StateTest from './stateTest/StateTest.vue';
 import WheelMenu from './wheelMenu/WheelMenu.vue';
+import BackgroundImageVue from './background/BackgroundImage.vue';
 
 export const CORE_IMPORTS = {
     Actions: shallowRef(Actions),
     Audio: shallowRef(Audio),
+    BackgroundImage: shallowRef(BackgroundImageVue),
     Designs: shallowRef(Designs),
     Icons: shallowRef(Icons),
     InputBox: shallowRef(InputBox),