+++ src/core/plugins/character-creator/webview/components/MenuButton.vue
@@ -1,6 +1,6 @@
 <template>
     <div
-        class="flex flex-row items-center justify-between transition-all cursor-pointer border border-neutral-700 bg-neutral-800 rounded-sm p-4 pt-2 pb-2 hover:text-neutral-200 hover:bg-neutral-700 active:scale-95"
+        class="flex flex-row items-center text-xs text-neutral-200 justify-between transition-all cursor-pointer border border-neutral-700 bg-neutral-800 rounded-sm p-4 pt-2 pb-2 hover:text-neutral-200 hover:bg-neutral-700 active:scale-95 h-full"
     >
         <slot></slot>
     </div>